test(security): Close mutation-testing gaps in auth/policy
cargo mutants (Gate G, #983) flagged several MISSED mutants in
crates/core/src/auth.rs and crates/core/src/policy.rs. Address them:

- ValidatedSecurityContext::new_for_scope's is_system-with-Unscoped
  fast path forces resolved_scope to System("all") two lines above a
  match on that same scope, making the Domain/Project arms of that
  match permanently unreachable. No test can kill a mutant in dead
  code, so simplify the block to a direct System lookup instead of
  matching on a value whose variant is already known.
- Add a test proving populate_user_domain actually populates the
  field (every existing fixture pre-populates user_domain, so the
  no-op branch was never exercised); the immediately-following
  validate() call depends on it.
- Add tests for PolicyEvaluationResult::valid() and the
  PolicyError -> KeystoneApiError conversion (only a genuine OPA
  Forbidden decision maps to 403; everything else must surface as
  500, per security.md), and for ValidatedSecurityContext's
  correlation_id() fallback.
- auth.rs:144's `expiry < Utc::now()` -> `<=` mutant would need a test
  asserting behavior at the exact instant expiry equals a live
  Utc::now() read, which requires clock injection to hit.
- SecurityContextTestingBuilder::expires_at is test-only fixture
  code; a no-op there breaks fixtures, not a production invariant.
- Credentials::try_from: assert the built roles list on the existing
  domain-scope test, killing the `!matches!(.., Unscoped)` guard
  being inverted (which would silently skip role population for
  every scoped request).
- PolicyEvaluationResult: add tests for the allow() accessor and the
  Display impl (which joins violation messages), neither of which
  had a direct test before.
- SecurityContext: add tests for the expires_at() getter, the
  set_token_restriction() setter (both previously only exercised
  through the private field or a builder, never the public API), and
  UserIdentityInfo::validate's user_id length boundary (64 valid, 65
  invalid, empty invalid) -- none of the three conditions on that
  check had a test forcing a specific pass/fail outcome before.
- PrincipalInfo: add tests for domain_id() and for validate()
  actually propagating a disabled-user failure through to the
  caller, rather than only exercising IdentityInfo::validate
  directly.
- Add SecurityContext::is_expired's `<` vs `<=` mutant to the
  exclude list alongside the existing app-cred expiry one: same
  clock-injection problem, not practically testable.
- resolve_system_roles's AuthenticationContext::Admin branch queries
  list_roles filtered to name "reader" instead of resolving the
  caller's real assignments, but no test ever constructed an Admin
  auth context to exercise it, so deleting that name filter (listing,
  and therefore granting, every role instead of just "reader") went
  undetected. Add a test asserting the query's name filter via
  MockRoleProvider::withf.
- SecurityContext::test_build's entire body already is
  SecurityContextTestingBuilder::default(), and that struct derives
  Default, so the reported mutant (replacing the body with
  Default::default()) is byte-for-byte equivalent to the original --
  add it to the exclude list with that rationale.
